day 25, two different approaches
I got the star just by feeding the graph to graphvis. with the
right layout engine (sfdp) it was clear which three edges
needed to be cut

then I decided to try something else. walk between a bunch of pairs
of random vertices and count traffic on each edge. the three "bridges"
consistently have the most traffic in the real input. they don't
in the example input, so I just pick the top 10 edges by traffic
and try each combination of 3 of them until I find one that works